Fail more gracefully if LyX isn't installed
If LyX isn't installed, LYXPATH will get set to "/bin/false", which will
cause the task to fail, but still doesn't give a satisfactory error
message. I suspect the real solution is to create a task whose only job
is to locate and store the path to LyX and make it a dependency of the
lyx2pdf task.